feat: add set and get pixel given coordinates
Since the pixel values are contained in a list of lists, this
wrapper methods allow someone to have more direct access
to the values. To ensure correctness, when someone asks for
(x,y) pixel, we return (y,x) value, since x is the horizontal position
(columns) and y is the vertical (lines).