feat(utils): Reading of different file structures
Since not all .pgm files will come in the same format this change is
needed to allow files not properly structure to work. Additional
validation has been implemented to ensure correctness of the file
and the Image object generated.
The option for passing the pixel data in the contructor has been
implemented.