fix: Fix 90 degrees rotation
The rotation was trying to use the same value list as the original
image, generating index out of range errors. This change
implements a "working copy" of the values after the rotation has
been done but before value atrtribution. Then the result is set to
the `self`object, concluding the operation. The Truth value of
`clockwise` had been inverted for this fix to work.